fix(geometry): split router/diagnostics.rs to absorb main's OpenTopologyRejected line
CI's Rust tests job runs against the auto-merge of this PR's head with the
current main tip, not the head commit alone. Between this PR's branch point
and now, main independently added a 3-line OpenTopologyRejected match arm to
router/diagnostics.rs (#3440/#3717), which this PR's diagnostics.rs (already
at its 586-line budget) does not have. The merged file lands at 589 lines,
tripping the module-size ratchet's grew-past-budget check even though this
PR's own branch and main's own tip each pass independently.

Split diagnostics.rs: the accumulator types (ClassificationStats,
HostOpeningDiagnostic, OpeningDiagnostic, OpeningKindDiag) and most of the
GeometryRouter recording methods move to a new sibling module,
router/diagnostics_recording.rs. record_host_failure_summary — the one
method whose body main's OpenTopologyRejected arm lands in — stays in
diagnostics.rs so a 3-way merge still applies main's addition as a normal
insertion instead of a delete/modify conflict (verified locally: merging
upstream/main into this branch now auto-merges cleanly and the ratchet
passes, where it previously either conflicted or landed at 589/586).

Both files land well under the 400-line ratchet limit, so diagnostics.rs's
allowlist row is deleted rather than raised, and the pinned rust/geometry
allowlist digest is updated to match.
